chore: renaming and moving snapshot loading
